containers/remoteLogger/sources/requirements.txt
psutil
docker
python-dotenv
mysql-connector-python
